pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-pipeline"
version = "0.1.0"
description = "Modular quantitative inference, risk, and execution pipeline"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Arian" }]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.5",
  "umap-learn>=0.5",
  "hdbscan>=0.8",
  "arch>=7.0",
  "pyyaml>=6.0",
  "typer>=0.12",
  "rich>=13.7",
  "matplotlib>=3.9",
  "plotly>=5.23",
  "streamlit>=1.37",
  "mlflow>=2.16",
  "requests>=2.32",
  "beautifulsoup4>=4.12",
  "pyarrow>=17.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "ruff>=0.6"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

